+++ packages/ag-charts-community/src/chart/navigator/navigator.ts
@@ -4,13 +4,13 @@ import type { ModuleContext } from '../../module/moduleContext';
 import type { BBox } from '../../scene/bbox';
 import type { Group } from '../../scene/group';
 import { setElementBBox } from '../../util/dom';
+import { initToolbarKeyNav } from '../../util/keynavUtil';
 import { Logger } from '../../util/logger';
 import { clamp, formatPercentage } from '../../util/number';
 import { ActionOnSet, ObserveChanges } from '../../util/proxy';
 import { AND, BOOLEAN, GREATER_THAN, LESS_THAN, OBJECT, POSITIVE_NUMBER, RATIO, Validate } from '../../util/validation';
 import { InteractionState, type PointerInteractionEvent } from '../interaction/interactionManager';
 import type { ZoomChangeEvent } from '../interaction/zoomManager';
-import { initToolbarKeyNav } from '../toolbar/toolbarUtil';
 import { RangeHandle } from './shapes/rangeHandle';
 import { RangeMask } from './shapes/rangeMask';
 import { RangeSelector } from './shapes/rangeSelector';
